requirements.txt
antlr4-python3-runtime==4.13.0
antlr4-tools==0.2
black==23.1.0
cfpq-data==4.0.2
pre-commit==3.0.4
pydot==1.4.2
pytest==7.2.1
scipy==1.10.1
